Reworked pwa hook into a context provider
