+++ src/app/hooks/index.ts
@@ -1,4 +1,3 @@
 export * from './api/api.hook';
 export * from './form/form.hook';
 export * from './local-storage/local-storage.hook';
-export * from './pwa/usePWA.hook';
