--- src/app/hooks/pwa/usePWA.hook.spec.tsx
@@ -1,9 +1,0 @@
-import { renderHook, act } from '@testing-library/react-hooks';
-import { useAddToHomescreenPrompt } from './usePWA.hook';
-
-describe('UseAddToHomescreenPrompt()', () => {
-    it('should change captured prompt', () => {
-        const { result } = renderHook(() => useAddToHomescreenPrompt());
-        expect(result);
-    });
-});
