+++ src/app/models/index.ts
@@ -6,3 +6,5 @@ export * from './Customer';
 export * from './Child';
 export * from './Parent';
 export * from './Cohort';
+
+export * from './BeforeInstall';
