+++ src/app/state/app/index.tsx
@@ -1,10 +1,16 @@
 import React from 'react';
 
 import { ErrorProvider } from './error/error.state';
+import { PWAProvider } from './pwa/pwa.state';
 
 const AppProvider: React.FC = ({ children }) => {
-    return <ErrorProvider>{children}</ErrorProvider>;
+    return (
+        <ErrorProvider>
+            <PWAProvider>{children}</PWAProvider>
+        </ErrorProvider>
+    );
 };
 
 export { AppProvider };
 export * from './error/error.state';
+export * from './pwa/pwa.state';
